Move Response to Responses::GetTracks
